Removed Pods Tracking + Intermittent changes
Removed Pods from the repo, that was stupid top trakc them to
begin with and I don't know why I did it. I also am just making a
commit here because it has been too long since I worked with the
repo and I just need to start somewhere.

+++ DrizzleTests/CurrentWeatherSpec.swift
@@ -57,8 +57,8 @@ class CurrentWeatherSpec: QuickSpec {
           
           expect(json["summary"].stringValue).to(equal("Partly Cloudy"))
           expect(json["icon"].stringValue).to(equal("partly-cloudy"))
-          expect(json["temperature"].doubleValue).to(equal(56))
-          expect(json["apparent_temp"].doubleValue).to(equal(50))
+          expect(json["temperature"].intValue).to(equal(56))
+          expect(json["apparent_temp"].intValue).to(equal(50))
         }
       }
     }
